# Conflicts:
#	rs-amberskynet/Cargo.lock

+++ rs-amberskynet/src/gfx/defines.rs
@@ -1,7 +1,7 @@
 use crate::core::{Array2D, Size2D};
 use crate::core_gfx::pixel::PixelRGBA;
 
-pub type AxeDimension = u32;
+pub type AxeDimension = usize;
 pub type ByteDimension = u8;
 
 pub type Size2d = Size2D<AxeDimension>;
